LaunchControlXL: had to revert to only 7 modes
The issue here was that the 8th factory mode is broken (different CCs than the 7 first ones) and the previous hack wasn't working well because it was impossible to get back to the mode 1.

We had to go back to 7 modes and I decided to make Send2Project available instead of Sends2Pan1 as this one seems more interesting to me.