Add MCU Extender Position Left to Right Preference
'The main unit will always be the most right unit and extenders will be added to the left.'

But Solid State Logic's SSL 360 is left to right.
If connected two or more UF8 and UF1, Will see channels 9-16 next 1-8. Very terrible.

This isn't Bitwig's fault, but I added it to resolve the issue like mine.

I have only tested up to 2x UF8. Not tested the 3, 4 units